# apps/api/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  draft.APIResponse:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      meta: {}
      success:
        type: boolean
    type: object
  draft.CreateDraftRequest:
    properties:
      article:
        type: string
      createdBy:
        type: string
      excerpt:
        type: string
      has_image:
        type: boolean
      image_prompt:
        type: string
      image_url:
        type: string
      keywords:
        items:
          type: string
        type: array
      scheduled_for:
        type: string
      seo_score:
        type: integer
      slug:
        type: string
      status:
        description: Tambahan yang dibutuhkan untuk update
        type: string
      target_products:
        items:
          type: string
        type: array
      teamID:
        type: string
      title:
        type: string
      topic:
        type: string
      update_at:
        type: string
      userID:
        type: string
    type: object
  draft.Draft:
    properties:
      article:
        type: string
      created_at:
        type: string
      created_by:
        type: string
      has_image:
        type: boolean
      id:
        type: string
      image_prompt:
        type: string
      image_url:
        type: string
      scheduled_for:
        type: string
      seo_score:
        type: integer
      status:
        type: string
      target_products:
        items:
          type: string
        type: array
      team_id:
        type: string
      title:
        type: string
      topic:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@seo-backend.com
    name: API Support
  description: API Documentation for SEO Backend Service
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: SEO Backend API
  version: "1.0"
paths:
...
  /api/drafts:
    get:
      consumes:
      - application/json
      description: Get all drafts with pagination and filters
      parameters:
      - description: 'Items per page (default: 10)'
        in: query
        name: limit
        type: integer
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: Search term
        in: query
        name: search
        type: string
      - description: Filter by status (draft, scheduled, published)
        enum:
        - draft
        - scheduled
        - published
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Drafts and stats
          schema:
            allOf:
            - $ref: '#/definitions/draft.APIResponse'
            - properties:
                data:
                  properties:
                    drafts:
                      type: object
                    stats:
                      type: object
                  type: object
              type: object
        "404":
          description: Drafts not found
          schema:
            $ref: '#/definitions/draft.APIResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/draft.APIResponse'
      security:
      - BearerAuth: []
      summary: Get all drafts
      tags:
      - Drafts
    post:
      consumes:
      - application/json
      description: Create a new draft article
      parameters:
      - description: Draft creation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/draft.CreateDraftRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Draft created
          schema:
            allOf:
            - $ref: '#/definitions/draft.APIResponse'
            - properties:
                data:
                  properties:
                    id:
                      type: string
                  type: object
              type: object
        "400":
          description: Invalid request or validation failed
          schema:
            $ref: '#/definitions/draft.APIResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/draft.APIResponse'
      security:
      - BearerAuth: []
      summary: Create a new draft
      tags:
      - Drafts
  /api/drafts/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a draft by ID
      parameters:
      - description: Draft ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Draft deleted
          schema:
            $ref: '#/definitions/draft.APIResponse'
        "404":
          description: Draft not found
          schema:
            $ref: '#/definitions/draft.APIResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/draft.APIResponse'
      security:
      - BearerAuth: []
      summary: Delete a draft
      tags:
      - Drafts
    get:
      consumes:
      - application/json
      description: Get a specific draft by its ID
      parameters:
      - description: Draft ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Draft details
          schema:
            allOf:
            - $ref: '#/definitions/draft.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/draft.Draft'
              type: object
        "404":
          description: Draft not found
          schema:
            $ref: '#/definitions/draft.APIResponse'
      security:
      - BearerAuth: []
      summary: Get draft by ID
      tags:
      - Drafts
    put:
      consumes:
      - application/json
      description: Update an existing draft by ID
      parameters:
      - description: Draft ID
        in: path
        name: id
        required: true
        type: string
      - description: Draft update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/draft.CreateDraftRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Draft updated
          schema:
            $ref: '#/definitions/draft.APIResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/draft.APIResponse'
        "404":
          description: Draft not found
          schema:
            $ref: '#/definitions/draft.APIResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/draft.APIResponse'
      security:
      - BearerAuth: []
      summary: Update a draft
      tags:
      - Drafts
  /api/drafts/{id}/publish:
    post:
      consumes:
      - application/json
      description: Publish a draft by ID
      parameters:
      - description: Draft ID
        in: path
        name: id
        required: true
        type: string
      - description: Publish request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/draft.CreateDraftRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Draft published
          schema:
            allOf:
            - $ref: '#/definitions/draft.APIResponse'
            - properties:
                data:
                  properties:
                    all_failed:
                      type: boolean
                    failed_count:
                      type: integer
                    published_at:
                      type: string
                    results:
                      type: array
                    some_failed:
                      type: boolean
                    status:
                      type: string
                    success_count:
                      type: integer
                    total_products:
                      type: integer
                  type: object
              type: object
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/draft.APIResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/draft.APIResponse'
      security:
      - BearerAuth: []
      summary: Publish a draft
      tags:
      - Drafts
  /api/drafts/{id}/reschedule:
    put:
      consumes:
      - application/json
      description: Reschedule a scheduled draft to a new date/time
      parameters:
      - description: Draft ID
        in: path
        name: id
        required: true
        type: string
      - description: 'Reschedule request (ISO 8601 format: 2026-07-07T15:30:00+07:00)'
        in: body
        name: request
        required: true
        schema:
          properties:
            scheduled_for:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Draft rescheduled
          schema:
            allOf:
            - $ref: '#/definitions/draft.APIResponse'
            - properties:
                data:
                  properties:
                    draft_id:
                      type: string
                    scheduled_for:
                      type: string
                    status:
                      type: string
                    updated_at:
                      type: string
                  type: object
              type: object
        "400":
          description: Invalid request or scheduled time must be in future
          schema:
            $ref: '#/definitions/draft.APIResponse'
        "404":
          description: Draft not found
          schema:
            $ref: '#/definitions/draft.APIResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/draft.APIResponse'
      security:
      - BearerAuth: []
      summary: Reschedule a draft
      tags:
      - Drafts
  /api/drafts/{id}/seo-score:
    get:
      consumes:
      - application/json
      description: Get SEO score for a draft
      parameters:
      - description: Draft ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: SEO score
          schema:
            allOf:
            - $ref: '#/definitions/draft.APIResponse'
            - properties:
                data:
                  type: object
              type: object
        "404":
          description: Draft not found
          schema:
            $ref: '#/definitions/draft.APIResponse'
      security:
      - BearerAuth: []
      summary: Get SEO score
      tags:
      - Drafts
  /api/drafts/{id}/similarity:
    get:
      consumes:
      - application/json
      description: Check similarity of a draft with other drafts
      parameters:
      - description: Draft ID
        in: path
        name: id
        required: true
        type: string
      - description: 'Items per page (default: 10)'
        in: query
        name: limit
        type: integer
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Similarity results
          schema:
            allOf:
            - $ref: '#/definitions/draft.APIResponse'
            - properties:
                data:
                  properties:
                    draft_id:
                      type: string
                    similar_drafts:
                      type: array
                    total:
                      type: integer
                  type: object
              type: object
        "404":
          description: Draft not found
          schema:
            $ref: '#/definitions/draft.APIResponse'
      security:
      - BearerAuth: []
      summary: Check similarity
      tags:
      - Drafts
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"